\name{Hex272}
\alias{Hex272}
\docType{data}
\title{
Global hexagonal grids
}
\description{
The points are produced by a hexagonal tessellation at a resolution of 272 grid locations to evaluate the point-pattern properties.
}
\usage{data(Hex272)}
\format{
  \describe{
    \item{\code{number}}{a numeric vector}
    \item{\code{longitude}}{a numeric vector}
    \item{\code{latitude}}{a numeric vector}
  }
}
\references{
Robeson, S.M., Li, A., Huang, C., 2014. Point-pattern analysis on the sphere. Spatial Statistics. 10, 76-86.
}
\keyword{datasets}
