\name{SphericalK-package}
\alias{SphericalK-package}
\alias{SphericalK}
\docType{package}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Spherical K-function
}
\description{
Spherical K-function for point-pattern analysis on the sphere
}
\details{
\tabular{ll}{
Package: \tab SphericalK\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-11-25\cr
License: \tab GPL-2 \cr
}
}
\author{
Scott Robeson, Ao Li, Chunfeng Huang

Maintainer: Ao Li<liao@umail.iu.edu>
}
\references{
Robeson, S.M., Li, A., Huang, C., 2014. Point-pattern analysis on the sphere. Spatial Statistics. 10, 76-86.
}
\keyword{ package }

